tests: Invoke generated code
Use `cargo script` for compilation as it supports building
without invocation, but it does require a main().

Use kscript for kotlin as it allows running .kt files
and can be used to add dependencies.

Remove build artefacts before and after compilation to
detect when an exe was produced.

Related to https://github.com/adsharma/py2many/issues/25