inference: use classes in type_map
Using strings in language specific type maps has the problem that it
doesn't cover the case where the type has been imported using an alias.

This is a step towards handling that correctly, although more work needs
to be done. We still have checks against FIXED_WIDTH_INTS_NAME in a few
places