rewriter: unpack scopes
Some rewriters that expand a single statement or expression to multiple
introduce new scopes. They do so because the visitor pattern has no way
of reaching out to the parent node. But this is undesirable because
sometimes we want the variables in the new scope to be available outside
the scope.

Using a separate pass contained in this rewriter, we can undo the new
scope introduced. We use the `is_block()` logic to detect such scopes.