analysis: fix the order of lookup
We need to search the scopes in reverse order. It exposed a second
bug. Which is that a variable defined in the outer scope is redefined
in the current scope according to the current logic in visit_Assign_one.

Fix the logic to define it only if it hasn't been defined before.

Fixes: https://github.com/adsharma/py2many/issues/417